Fix memory leak in filename_complete_indir function
Fix memory leak when completing in a deleted directory by ensuring os_findclose is called regardless of the opened status.

I was working on my minishell project for 42 Barcelona and identified this bug using Valgrind and fsanitize. 

I also want to say thank you for creating this tool; it has been incredibly helpful for my project.